Move functions to PSF utils
